chore: follow best preatice
